requirements.txt
python-whois==0.8.0
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.16.0
